backend/requirements.txt
Flask
Werkzeug
python-dotenv
pyjwt
flask_mail